Feat: Card 컴포넌트 반응형 완성
