+++ pages/test/index.tsx
@@ -1,11 +1,9 @@
-import Card from '@/src/components/ui/Card';
-
-// import mockupData from './mockup.json';
+import CardStructure from '@/src/components/ui/Card/CardStructure';
 
 export default function Page() {
 	return (
 		<div className='flex flex-col gap-6 p-1'>
-			<Card />
+			<CardStructure />
 		</div>
 	);
 }
